♻️  conversation service refactor
