+++ backend/services/conversation_management_service.py
@@ -1,12 +1,17 @@
 import json
 import logging
-from typing import List, Optional
+import os
+from typing import List, Optional, Dict, Any
+from datetime import datetime
 
 from fastapi import HTTPException, Header
+from smolagents import OpenAIServerModel
 
 from consts.model import MessageRequest, ConversationResponse, AgentRequest, MessageUnit
 from database.conversation_db import create_conversation_message, create_source_search, create_message_units, \
-    create_source_image
+    create_source_image, rename_conversation, get_conversation_list, get_conversation_history, get_source_images_by_message, \
+    get_source_images_by_conversation, get_source_searches_by_message, get_source_searches_by_conversation, \
+    delete_conversation, get_conversation, create_conversation, update_message_opinion
 
 
 logger = logging.getLogger("conversation_management_service")
@@ -144,20 +149,15 @@ def save_message(request: MessageRequest, authorization: Optional[str] = Header(
 
 
 def save_conversation_user(request: AgentRequest, authorization: Optional[str] = None):
-    user_role_count = count_user(request)
+    user_role_count = sum(1 for item in getattr(request, "history", []) if item.get("role") == "user")
 
     conversation_req = MessageRequest(conversation_id=request.conversation_id, message_idx=user_role_count * 2,
         role="user", message=[MessageUnit(type="string", content=request.query)], minio_files=request.minio_files)
     save_message(conversation_req, authorization=authorization)
 
 
-def count_user(request):
-    # Use generator expression and sum function to count the number of roles that are user
-    return sum(1 for item in getattr(request, "history", []) if item.get("role") == "user")
-
-
 def save_conversation_assistant(request: AgentRequest, messages: List[str], authorization: Optional[str] = None):
-    user_role_count = count_user(request)
+    user_role_count = sum(1 for item in getattr(request, "history", []) if item.get("role") == "user")
 
     message_list = []
     for item in messages:
@@ -170,3 +170,413 @@ def save_conversation_assistant(request: AgentRequest, messages: List[str], auth
     conversation_req = MessageRequest(conversation_id=request.conversation_id, message_idx=user_role_count * 2 + 1,
         role="assistant", message=message_list, minio_files=request.minio_files)
     save_message(conversation_req, authorization=authorization)
+
+
+def extract_user_messages(history: List[Dict[str, str]]) -> str:
+    """
+    Extract user message content from conversation history
+
+    Args:
+        history: List of conversation history records
+
+    Returns:
+        str: Concatenated user message content
+    """
+    content = ""
+    for message in history:
+        if message.get("role") == "user" and message.get("content"):
+            content += f"\n### User Question：\n{message['content']}\n"
+        if message.get("role") == "assistant" and message.get("content"):
+            content += f"\n### Response Content：\n{message['content']}\n"
+    return content
+
+
+def call_llm_for_title(content: str) -> str:
+    """
+    Call LLM to generate a title
+
+    Args:
+        content: Conversation content
+
+    Returns:
+        str: Generated title
+    """
+    # Create OpenAIServerModel instance
+    llm = OpenAIServerModel(model_id=os.getenv('LLM_MODEL_NAME'), api_base=os.getenv('LLM_MODEL_URL'),
+        api_key=os.getenv('LLM_API_KEY'), temperature=0.7, top_p=0.95)
+
+    # Build messages
+    messages = [{"role": "system",
+                 "content": "Please generate a concise and accurate title (no more than 12 characters) for the following conversation, highlighting the core theme or key information. The title should be natural and fluent, avoiding forced keyword stacking.\n\n**Title Generation Requirements:**\n1. If the conversation involves specific questions, the title should summarize the essence of the question (e.g., 'How to solve XX issue?');\n2. If the conversation is about knowledge sharing, the title should highlight the core knowledge point (e.g., 'Key Steps of Photosynthesis');\n3. Avoid using generic terms like 'Q&A' or 'Consultation', prioritize domain specificity;\n4. The title language should match the conversation language.\n\n**Examples:**\n- Conversation: User asking about multiple methods for Python list deduplication → Title: Python List Deduplication Techniques\n- Conversation: Discussion about factors affecting new energy vehicle battery life → Title: Factors Affecting EV Battery Life\n- Conversation: 你好 → Title: 你好\n\nPlease output only the generated title without additional explanation."},
+        {"role": "user", "content": f"Please generate a concise title (no more than 12 characters) based on the following conversation:\n{content}\n\nTitle："}]
+
+    # Call the model
+    response = llm(messages, max_tokens=10)
+
+    return response.content.strip()
+
+
+def update_conversation_title(conversation_id: int, title: str, user_id: str = None) -> bool:
+    """
+    Update conversation title
+
+    Args:
+        conversation_id: Conversation ID
+        title: New title
+        user_id: Reserved parameter, user ID
+    Returns:
+        bool: Whether the update was successful
+    """
+    success = rename_conversation(conversation_id, title)
+    if not success:
+        raise HTTPException(status_code=404, detail=f"Conversation {conversation_id} does not exist or has been deleted")
+    return success
+
+
+def create_new_conversation(title: str) -> Dict[str, Any]:
+    """
+    Create a new conversation
+
+    Args:
+        title: Conversation title
+
+    Returns:
+        Dict containing conversation data
+    """
+    try:
+        conversation_data = create_conversation(title)
+        return conversation_data
+    except Exception as e:
+        logging.error(f"Failed to create conversation: {str(e)}")
+        raise HTTPException(status_code=500, detail=str(e))
+
+
+def get_conversation_list_service() -> List[Dict[str, Any]]:
+    """
+    Get all conversation list
+
+    Returns:
+        List of conversation data
+    """
+    try:
+        conversations = get_conversation_list()
+        return conversations
+    except Exception as e:
+        logging.error(f"Failed to get conversation list: {str(e)}")
+        raise HTTPException(status_code=500, detail=str(e))
+
+
+def rename_conversation_service(conversation_id: int, name: str) -> bool:
+    """
+    Rename a conversation
+
+    Args:
+        conversation_id: Conversation ID
+        name: New conversation title
+
+    Returns:
+        bool: Whether the rename was successful
+    """
+    try:
+        success = rename_conversation(conversation_id, name)
+        if not success:
+            raise HTTPException(
+                status_code=404,
+                detail=f"Conversation {conversation_id} does not exist or has been deleted"
+            )
+        return True
+    except Exception as e:
+        logging.error(f"Failed to rename conversation: {str(e)}")
+        if isinstance(e, HTTPException):
+            raise e
+        raise HTTPException(status_code=500, detail=str(e))
+
+
+def delete_conversation_service(conversation_id: int) -> bool:
+    """
+    Delete specified conversation
+
+    Args:
+        conversation_id: Conversation ID to delete
+
+    Returns:
+        bool: Whether the deletion was successful
+    """
+    try:
+        success = delete_conversation(conversation_id)
+        if not success:
+            raise HTTPException(
+                status_code=404,
+                detail=f"Conversation {conversation_id} does not exist or has been deleted"
+            )
+        return True
+    except Exception as e:
+        logging.error(f"Failed to delete conversation: {str(e)}")
+        if isinstance(e, HTTPException):
+            raise e
+        raise HTTPException(status_code=500, detail=str(e))
+
+
+def get_conversation_history_service(conversation_id: int) -> Dict[str, Any]:
+    """
+    Get complete history of specified conversation
+
+    Args:
+        conversation_id: Conversation ID
+
+    Returns:
+        Dict containing conversation history data
+    """
+    try:
+        # Get original conversation history data
+        history_data = get_conversation_history(conversation_id)
+
+        if not history_data:
+            raise HTTPException(
+                status_code=404,
+                detail=f"Conversation {conversation_id} does not exist or has been deleted"
+            )
+
+        # Collect search content, grouped by message_id
+        search_by_message = {}
+        for record in history_data['search_records']:
+            message_id = record['message_id']
+            # Process published_date, ensure it's a datetime object
+            published_date = None
+            if record['published_date'] is not None:
+                if isinstance(record['published_date'], datetime):
+                    published_date = record['published_date'].strftime("%Y-%m-%d")
+                elif isinstance(record['published_date'], str):
+                    published_date = record['published_date']
+
+            # Build search content
+            search_item = {"title": record["source_title"], "text": record["source_content"],
+                           "source_type": record["source_type"], "url": record["source_location"],
+                           "filename": record["source_title"] if record["source_type"] == "file" else None,
+                           "published_date": published_date, "score": record["score_overall"],
+                           "cite_index": record["cite_index"], "search_type": record["search_type"],
+                           "tool_sign": record["tool_sign"], "score_details": {}}
+
+            if record["score_accuracy"] is not None:
+                search_item["score_details"]["accuracy"] = record["score_accuracy"]
+            if record["score_semantic"] is not None:
+                search_item["score_details"]["semantic"] = record["score_semantic"]
+
+            if message_id not in search_by_message:
+                search_by_message[message_id] = []
+            search_by_message[message_id].append(search_item)
+
+        # Collect image content - grouped by message_id
+        image_by_message = {}
+        for record in history_data['image_records']:
+            message_id = record['message_id']
+            if message_id not in image_by_message:
+                image_by_message[message_id] = []
+            image_by_message[message_id].append(record['image_url'])
+
+        # Sort by message index and build final message list, including images and search content
+        messages = []
+
+        for msg in history_data['message_records']:
+            message_id = msg['message_id']
+            role = msg['role']
+            message_content = msg['message_content']
+
+            if role == 'user':
+                # User message: directly use message_content as message field value
+                message_item = {
+                    'role': role,
+                    'message': message_content,
+                    'message_id': message_id,
+                    'opinion_flag': None
+                }
+
+                # Add minio_files field (if any)
+                if 'minio_files' in msg and msg['minio_files']:
+                    message_item['minio_files'] = msg['minio_files']
+            else:
+                # Assistant message: message is an array, need to add final_answer type message unit
+                message_units = msg['units'] or []
+                # Add final_answer type message unit
+                message_units.append({
+                    'type': 'final_answer',
+                    'content': message_content
+                })
+
+                message_item = {
+                    'role': role,
+                    'message': message_units,
+                    'message_id': message_id,
+                    'opinion_flag': msg['opinion_flag']
+                }
+
+            # Add image content (if any)
+            if message_id in image_by_message:
+                message_item['picture'] = image_by_message[message_id]
+
+            # Add search content (if any)
+            if message_id in search_by_message:
+                message_item['search'] = search_by_message[message_id]
+
+            messages.append(message_item)
+
+        # Build final result
+        formatted_history = {
+            'conversation_id': str(history_data['conversation_id']),  # Convert to string
+            'create_time': history_data['create_time'],
+            'message': messages
+        }
+
+        return [formatted_history]  # Wrap in list to match expected format
+
+    except Exception as e:
+        logging.error(f"Failed to get conversation history: {str(e)}")
+        if isinstance(e, HTTPException):
+            raise e
+        raise HTTPException(status_code=500, detail=str(e))
+
+
+def get_sources_service(conversation_id: Optional[int], message_id: Optional[int], source_type: str = "all") -> Dict[str, Any]:
+    """
+    Get message source information (images and search results)
+
+    Args:
+        conversation_id: Optional conversation ID
+        message_id: Optional message ID
+        source_type: Source type, default is "all", options are "image", "search", or "all"
+
+    Returns:
+        Dict containing source information
+    """
+    try:
+        if not conversation_id and not message_id:
+            return {
+                "code": 400,
+                "message": "Must provide conversation_id or message_id parameter",
+                "data": None
+            }
+
+        # If conversation ID is provided
+        if conversation_id:
+            conversation = get_conversation(conversation_id)
+            if not conversation:
+                return {
+                    "code": 404,
+                    "message": f"Conversation {conversation_id} does not exist",
+                    "data": None
+                }
+
+        result = {"searches": [], "images": []}
+
+        # Get image sources
+        if source_type in ["image", "all"]:
+            images = []
+            if message_id:
+                image_records = get_source_images_by_message(message_id)
+            elif conversation_id:
+                image_records = get_source_images_by_conversation(conversation_id)
+
+            for image in image_records:
+                images.append(image["image_url"])
+
+            result["images"] = images
+
+        # Get search sources
+        if source_type in ["search", "all"]:
+            searches = []
+            if message_id:
+                search_records = get_source_searches_by_message(message_id)
+            elif conversation_id:
+                search_records = get_source_searches_by_conversation(conversation_id)
+
+            for record in search_records:
+                search_item = {
+                    "title": record["source_title"],
+                    "text": record["source_content"],
+                    "source_type": record["source_type"],
+                    "url": record["source_location"],
+                    "filename": record["source_title"] if record["source_type"] == "file" else None,
+                    "published_date": record["published_date"].strftime("%Y-%m-%d") if record[
+                        "published_date"] else None,
+                    "score": record["score_overall"]
+                }
+
+                search_item["score_details"] = {}
+                if record["score_accuracy"] is not None:
+                    search_item["score_details"]["accuracy"] = record["score_accuracy"]
+                if record["score_semantic"] is not None:
+                    search_item["score_details"]["semantic"] = record["score_semantic"]
+
+                if conversation_id and not message_id:
+                    search_item["message_id"] = record["message_id"]
+
+                searches.append(search_item)
+
+            result["searches"] = searches
+
+        return {
+            "code": 0,
+            "message": "success",
+            "data": result
+        }
+
+    except Exception as e:
+        logging.error(f"Failed to get message sources: {str(e)}")
+        return {
+            "code": 500,
+            "message": str(e),
+            "data": None
+        }
+
+
+def generate_conversation_title_service(conversation_id: int, history: List[Dict[str, str]]) -> str:
+    """
+    Generate conversation title
+
+    Args:
+        conversation_id: Conversation ID
+        history: Conversation history list
+
+    Returns:
+        str: Generated title
+    """
+    try:
+        # Extract user messages
+        content = extract_user_messages(history)
+
+        # Call LLM to generate title
+        title = call_llm_for_title(content)
+
+        # Update conversation title
+        update_conversation_title(conversation_id, title)
+
+        return title
+
+    except Exception as e:
+        logging.error(f"Failed to generate conversation title: {str(e)}")
+        if isinstance(e, HTTPException):
+            raise e
+        raise HTTPException(status_code=500, detail=str(e))
+
+
+def update_message_opinion_service(message_id: int, opinion: Optional[str]) -> bool:
+    """
+    Update message like/dislike status
+
+    Args:
+        message_id: Message ID
+        opinion: Opinion value ('Y' or 'N' or None)
+
+    Returns:
+        bool: Whether the update was successful
+    """
+    try:
+        success = update_message_opinion(message_id, opinion)
+        if not success:
+            raise HTTPException(status_code=404, detail="Message does not exist or has been deleted")
+        return True
+    except Exception as e:
+        logging.error(f"Failed to update message like/dislike: {str(e)}")
+        if isinstance(e, HTTPException):
+            raise e
+        raise HTTPException(status_code=500, detail=str(e))
